requirements.txt
python-dotenv==1.0.1
langchain==0.2.1
langchain-community==0.2.1
langchain-google-genai==1.0.5
bs4==0.0.2
langchainhub==0.1.18
wikipedia==1.4.0
tavily-python==0.3.3
wikipedia
sympy
